Restructure app to avoid tar/untar steps.
Note: This merges the /app and /omeka directories from
previous versions. Since /app was a bind mount and /omeka
was not, I have removed the bind mount to avoid potential
conflicts.

I am intending to replace the inotify-based code reloading
with one based on Docker compose watch, but at this commit,
neither reload method is in place.